# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Otus API Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exercises": {
            "get": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Get list of exercises",
                "description": "Returns list of exercises",
                "operationId": "getExercises",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExerciseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Store new exercise",
                "description": "Returns new created exercise",
                "operationId": "storeExercise",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreExerciseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExerciseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error"
                    }
                }
            }
        },
        "/api/exercises/{id}": {
            "get": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Get exercise information",
                "description": "Returns exercise data",
                "operationId": "getExerciseById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Exercise id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExerciseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Update exercise",
                "description": "Returns updated exercise",
                "operationId": "updateExercise",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateExerciseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExerciseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Delete existing exercise",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteExercise",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Exercise id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreExerciseRequest": {
                "title": "Store exercise request",
                "description": "Store exercise request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "example": "Test"
                    }
                },
                "type": "object"
            },
            "UpdateExerciseRequest": {
                "title": "Update exercise request",
                "description": "Store exercise request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "example": "Test some"
                    }
                },
                "type": "object"
            },
...
            "ExerciseResource": {
                "title": "ExerciseResource",
                "description": "Exercise resource",
                "properties": {
                    "name": {
                        "title": "name",
                        "example": "Ex. #3y7-2y3"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ExerciseResource"
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}